+++ CHANGELOG.md
@@ -1,1 +1,15 @@
-v1.1.0 [FEATURE] Added Employees in Employee Management in Admin Dashboard, Added Shift Schedules and Leave Types at User Dashboard, CHANGELOG, LICENSE, README
+v1.1.2 [FEATURE]
+        -- Roles
+            -- Added Roles for Access Level Functionality in Employee Account Creation
+            -- Created 'user_roles' table
+        -- Login
+            -- Login Page connected at the ClockWise Database
+            -- Login Redirect based on Username assigned Role
+        -- Session
+            -- Placed session_start() at config file
+        
+v1.1.1 [FEATURE]
+        -- Add New Employees Functionality
+
+v1.1.0 [FEATURE]
+        -- Added Employees in Employee Management in Admin Dashboard, Added Shift Schedules and Leave Types at User Dashboard, CHANGELOG, LICENSE, README